[doc] Drop some unneeded text from boot rom document
This drops the document organization section from boot rom document.
That was probably needed in the initial RFC, but isn't really needed
now that this is a manual/document.

Similarly, we drop some in-document layout text. This comes from when
the text was a proposal, rather than a document about what exists. We
probably don't need the structure notes in the document's new form.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>