[bazel/fpga] Split off VCU118 bitstream generation from CW340
When doing *sw* tests on the FPGA, don't reload the bitstream each
time to avoid the long reconfiguration time (we also assume the
bitstream is already loaded).

I've decided to keep the rules in //hw/bitstream/vcu118 consistent
with the other boards in //hw/bitstream, but the consequence is
that you must pass `--define bitstream=skip` on software tests.

Signed-off-by: Quan Nguyen <qmn@zerorisc.com>