[fpga] Add bitstream loading to VCU118 execution environment
Introduces a new VCU118Flavor for the hyperdebug transport type in opentitanlib.
It uses the program_vcu118.sh/tcl scripts for bitstream loading which have been added as embedded data.

Additionally we add a vcu118.json configuration file.
This is a copy of the hyperdebug_cw340.json file with only the interface name changed.
The copy is necessary because opentitanlib checks if the interface name and flavor name match.

Example Usage:
```
./bazelisk.sh run //sw/host/opentitantool -- --logging debug --interface vcu118 fpga load-bitstream /home/bitstreams/lowrisc_systems_chip_earlgrey_vcu118_0.1.bit
```

Closes #53

Signed-off-by: Martin Troiber <troi@zerorisc.com>