[pinmux,dv] Use multireg_idx in pinmux.hjson
This overrides the reset value of one copy of the mio_pad_attr
register to reflect its differing reset value.

We also change the "tags" for that register to turn its reset value
test back on (since the config file now gives the reset value we
expect).

Note that this isn't a very satisfying solution: the magic "30" index
comes from a TargetCfg parameter, which is passed in from a chip
top-level PinmuxTargetCfg parameter.

The nicest design would figure out the index as it instantiated
hw/top_earlgrey/.../autogen/pinmux.hjson. This index fundamentally
comes from topgen's merge.py, but the path for that index is somewhat
complicated.

Another reasonable (if not "nicest") design would not set a special
index in pinmux.hjson.tpl but would instead tweak the RAL code so that
it used set_reset() on the correct signal by looking at a parameter in
the UVM build phase. That is a bit ugly because we wouldn't end up
with nice documentation for the register and we'd have to bodge some
information about pinmux into each top-level testbench.

This commit does the laziest approach, where we hardcode the magic
"30" into pinmux.hjson. This will work fine until we have two
different top-levels where this special index happens to be different.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>