[sival] use different region and page in owner stage for flash_ctrl_test
Currently it always use region 0 and first page of bank 1, but both of
them are occupied by ROM_EXT. Simply switch to region 2 and page 0x20
when running in owner stage would make the test to work on owner stage.

Existing checks for silicon is wrong -- it should be checking for boot
stages instead. These are replaced too.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>