copy over the HJSON spec for the PQ OTBN
This is a full copy of the HJSON file from the PQ working repo.

This contains references to a kmac block which are not actually
incorporated in this patch set (it is in the simulator but not
in the hardware).

But more importantly, it also expands the size of the IMEM/DMEM
for OTBN, which is necessary in order for the "assert" to pass
on the size check in the dmem compare.