modify the memory comparison to match the outputs of the routines
More API divergence. it looks like what might be happening is some
fairly fancy provision was baked into the expo main branch which
can go into an ELF file and extract multiple disjoint regions
of memory for comparison.

This would imply that the data is labeled and tagged for such
comparisons, but no such labels exist in the actual test data
provided.

This change *probably* breaks some other tests that depend on
this feature, so while this allows us to run this one PQ test
case, I have a strong feeling that this is the wrong approach.

However, it's easy to pull out commits so let's drop this
into a PR and see what reviewers have to say.