convert_exp.py is a tool for converting .exp files from pq-opentitan
Usage:

1. Copy the routine from pq-opentitan into the desired location and setup
the BUILD template

2. Copy the test from pq-opentitan into the desired location, setup BUILD,
*but* in this case, rename the '.exp' file with an extension of '.exp2'.
The '.exp2' file serves as a reference to generate .exp/.dexp files.

3. Run the script using the test target you set up, e.g.:

`convert_exp.py --test_target //sw/otbn/crypto/tests/mldsa:poly_add_dilithium_test`

This will attempt to do the following:

1. Build the test, to create a .elf file by trying to run the test. It's
OK if the test fails, but we do need a .elf file.

2. From the .elf file, dumps all the symbols with `nm`

3. Tries to match based on address in dmem any contiguous memory regions
in the '.exp2' file, and assigns them to the label corresponding to
that label in the code. The matched areas are written to a '.dexp' file
that is consistent with the current format in expo.

4. Also checks to see if register values are specified, if the line
pattern-matches against a register spec, it will also output a '.exp'
file that contains the expceted register contents.

Regions that can't be matched are printed as an error.

Signed-off-by: bunnie <bunnie@kosagi.com>