[otbn] Fix copyright headers on a couple of files.
While checking the headers, I noticed that these failed because of an outdated
lowRISC copyright format. However, on closer inspection, the files seem to be
entirely new content (which is why the earlier rebase didn't add the new
format) and have a "Copyright" line instead of a "Modified" line on them. My
best understanding is that this is a mistake and the "authors" copyright was
supposed to go on top like it does for the other new files, so I'm correcting
it here according to that understanding.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>