[dv] Tidy up and document cip_base_scoreboard.predict_tl_err
The previous version of the documentation was written by me in 2021
and I've learned a bit more about both the DV environment and UVM in
the meantime.

The changes to the code are just rearranging things: there should be
no functional change.

Also, make the function protected: it's only used from scoreboards, so
doesn't need to be public.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>