[entropy_src] Lock up Ack SM upon fatal local events, align scoreboard
This commit modifies the RTL to move the Ack SM into the terminal error
state whenever a fatal local event is observed. Previously, the Ack SM
would only move to this state upon observing one particular fatal error
event, namely the counter error event. As the fatal alert was signaled
already before, this is not a very critical issue. But the new behavior
is more desirable as it means the ENTROPY_SRC immediately stops serving entropy whenever
a fatal local event is observed.

In addition, the modeling in the scoreboard is re-aligned with the
actual RTL implementation which fixes a notable amount of test failures.

This resolves lowRISC/OpenTitan#24410.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>