[otbn] Fixed dvsim errors from snippet generators due to increased IMEM/DMEM size
Fixed oversized immediate value from bad_load_store due to a DMEM size > 16KB
Added error handling for bad_deep_loop to prevent a build failure. If IMEM is
greater than 16KB there is a significant chance that the loop does not hit a
memory address that has already been written to. It is also impossible for the
bodysize to reach the last address in memory, without the current PC being within
range, therefore a None path has been added.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>