[otbn] Fixed bug in ctrl_redun vseq causing improper signal sampling
The ctrl_redun test injects errors to OTBN MAC and bignum ALU
and sends the error to the ISS model. Much of the injection relies
on determining when certain components are active using the
uvm_hdl_data tasks. Sampling did not include a width mask and
was leading to X in the signal for some reads in VCS/Verdi.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>